--- chiizu-lib-release/build.gradle
@@ -1,2 +1,0 @@
-configurations.create("default")
-artifacts.add("default", file('chiizu-lib-release.aar'))
