--- chiizu-lib/src/main/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="chiizu">
-
-    <application android:label="@string/app_name"
-        android:supportsRtl="true">
-
-    </application>
-
-</manifest>
