--- lib/assets/ChiizufileTemplate
@@ -1,15 +1,0 @@
-# remove the leading '#' to uncomment lines
-
-# app_package_name 'your.app.package'
-# use_tests_in_packages ['your.screenshot.tests.package']
-
-# app_apk_path 'path/to/your/app.apk'
-# tests_apk_path 'path/to/your/tests.apk'
-
-locales ['en-US', 'fr-FR', 'it-IT']
-
-# clear all previously generated screenshots in your local output directory before creating new ones
-clear_previous_screenshots
-
-# For more information about all available options run
-#   chiizu --help
