--- lib/chiizu/version.rb
@@ -1,4 +1,0 @@
-module Chiizu
-  VERSION = "1.0.0".freeze
-  DESCRIPTION = "Automate taking localized screenshots of your Android app on real devices or emulators".freeze
-end
