+++ lib/screengrab.rb
@@ -1,0 +1,30 @@
+require 'fastlane_core'
+
+require 'screengrab/version'
+require 'screengrab/runner'
+require 'screengrab/detect_values'
+require 'screengrab/dependency_checker'
+require 'screengrab/options'
+require 'screengrab/android_environment'
+
+require 'open3'
+
+module Screengrab
+  # Use this to just setup the configuration attribute and set it later somewhere else
+  class << self
+    attr_accessor :config
+    attr_accessor :android_environment
+
+    def config=(value)
+      @config = value
+      DetectValues.set_additional_default_values
+    end
+
+    def screengrabfile_name
+      "Screengrabfile"
+    end
+  end
+
+  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
+  UI = FastlaneCore::UI
+end
