+++ lib/screengrab/version.rb
@@ -1,0 +1,4 @@
+module Screengrab
+  VERSION = "1.0.0".freeze
+  DESCRIPTION = "Automate taking localized screenshots of your Android app on real devices or emulators".freeze
+end
