+++ screengrab-lib-release/build.gradle
@@ -1,0 +1,2 @@
+configurations.create("default")
+artifacts.add("default", file('screengrab-lib-release.aar'))
