+++ screengrab-lib/src/main/AndroidManifest.xml
@@ -1,0 +1,9 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="screengrab">
+
+    <application android:label="@string/app_name"
+        android:supportsRtl="true">
+
+    </application>
+
+</manifest>
